refactor(components): 컴포넌트 TS 마이그래이션
